refactored a bit

+++ src/js/Class.cpp
@@ -150,12 +150,12 @@ namespace DarkDescent::JS
 
 	v8::MaybeLocal<v8::Value> Class::instantiate(const std::vector<v8::Local<v8::Value>>& args) const
 	{
-		return getClass()->CallAsConstructor(env.context(), args.size(), const_cast<v8::Local<v8::Value>*>(args.data()));
+		return getClass()->CallAsConstructor(env.context(), static_cast<int>(args.size()), const_cast<v8::Local<v8::Value>*>(args.data()));
 	}
 
 	v8::MaybeLocal<v8::Value> Class::instantiate(std::vector<v8::Local<v8::Value>>&& args) const
 	{
-		return getClass()->CallAsConstructor(env.context(), args.size(), args.data());
+		return getClass()->CallAsConstructor(env.context(), static_cast<int>(args.size()), args.data());
 	}
